backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
celery
redis
django-cors-headers
Pillow
python-dotenv
# psycopg2-binary
